+++ htdocs/reciepe.php
@@ -120,7 +120,7 @@ if($item['reciepe']=='<no_fridge />'){
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
-<script type="text/javascript" src="scripts/custom.js"></script>
+<script type="text/javascript" src="scripts/custom.js?<?=$html_NoCache_Version?>"></script>
 
 <?php
 output_session_error();