+++ htdocs/upload.php
@@ -186,7 +186,7 @@ if($lang == 'en'){
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
-<script type="text/javascript" src="scripts/custom.js"></script>
+<script type="text/javascript" src="scripts/custom.js?<?=$html_NoCache_Version?>"></script>
 <script>
 document.addEventListener('DOMContentLoaded', function() {
   var fileInput = document.getElementById('file1_inp');